refactor: move connect cube out of common
