+++ shell/app/common/__tests__/components/connect.test.tsx
@@ -12,7 +12,7 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 import React from 'react';
-import { connectCube } from 'common';
+import { connectCube } from 'common/utils';
 import { mount } from 'enzyme';
 
 const Comp = (props = {}) => {
