+++ shell/app/common/utils/connect-cube.tsx
