+++ shell/app/common/utils/index.ts
@@ -19,6 +19,7 @@ import moment from 'moment';
 import { Key, pathToRegexp, compile } from 'path-to-regexp';
 import { AxiosResponse } from 'axios';
 
+export { connectCube } from './connect-cube';
 export { getBrowserInfo, getCookies, getLS, removeLS, setLS, clearLS, LSObserver } from './browser';
 export { replaceEmoji } from './emoji';
 export { goTo, pages, resolvePath } from './go-to';
