+++ shell/app/modules/application/common/app-list-protocol.tsx
@@ -20,8 +20,8 @@ import React from 'react';
 import { Spin } from 'core/nusi';
 import { useEffectOnce } from 'react-use';
 import i18n from 'i18n';
-import { Icon as CustomIcon, IF, connectCube } from 'common';
-import { goTo, ossImg } from 'common/utils';
+import { Icon as CustomIcon, IF } from 'common';
+import { goTo, ossImg, connectCube } from 'common/utils';
 import { theme } from 'app/themes';
 import { modeOptions, appMode } from 'application/common/config';
 import DiceConfigPage from 'config-page/index';
