+++ shell/app/modules/dcos/pages/purchase-cluster/purchase-cluster.tsx
@@ -14,8 +14,8 @@
 import { cloneDeep } from 'lodash';
 import i18n from 'i18n';
 import { Form, Steps, Button, Tooltip, Row, Col, Spin, Switch } from 'core/nusi';
-import { RenderPureForm, KeyValueList, Icon as CustomIcon, RenderFormItem, connectCube } from 'common';
-import { goTo } from 'common/utils';
+import { RenderPureForm, KeyValueList, Icon as CustomIcon, RenderFormItem } from 'common';
+import { goTo, connectCube } from 'common/utils';
 import React from 'react';
 import {
   supportLBRegion,
