+++ shell/app/modules/dcos/pages/service-manager/index.jsx
@@ -15,7 +15,8 @@ import React from 'react';
 import i18n from 'i18n';
 import { Breadcrumb, Spin, Select, Input } from 'core/nusi';
 import { cloneDeep, map, isEmpty, debounce } from 'lodash';
-import { JsonChecker, IF, Holder, connectCube } from 'common';
+import { JsonChecker, IF, Holder } from 'common';
+import { connectCube } from 'common/utils';
 import PureServiceList from './service-list';
 import AssociatedAddons from '../associated-addon';
 import { Right as IconRight } from '@icon-park/react';
