+++ shell/app/modules/publisher/pages/publisher-manage/publisher-list-v2.tsx
@@ -15,8 +15,8 @@ import React from 'react';
 import { Input, Spin, Button, Table } from 'core/nusi';
 import { useDebounce, useUnmount, useUpdateEffect } from 'react-use';
 import i18n from 'i18n';
-import { connectCube, useUpdate, CustomFilter } from 'common';
-import { goTo, insertWhen } from 'common/utils';
+import { useUpdate, CustomFilter } from 'common';
+import { connectCube, goTo, insertWhen } from 'common/utils';
 import './publisher-list-v2.scss';
 import routeInfoStore from 'core/stores/route';
 import ArtifactsFormModal from 'publisher/pages/artifacts/artifacts-form-modal';
